% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jma.R
\name{body_read_jma}
\alias{body_read_jma}
\title{generate body for POST}
\usage{
body_read_jma(from, to, phpsessid)
}
\arguments{
\item{from}{Date}

\item{to}{Date}

\item{phpsessid}{php session id generated by get_phpsessionid()}
}
\description{
generate body for POST
}
