% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_list.R
\name{get_stations_of_pd}
\alias{get_stations_of_pd}
\title{get list of stations in a prefecture}
\usage{
get_stations_of_pd(pd)
}
\arguments{
\item{pd}{String representing each prefecture}
}
\description{
get list of stations in a prefecture
}
