% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_list.R
\name{is_stmark}
\alias{is_stmark}
\title{predicate function to determin whether a node has class "stmark"}
\usage{
is_stmark(node)
}
\arguments{
\item{node}{xml node containg div.station}
}
\description{
predicate function to determin whether a node has class "stmark"
}
