% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_list.R
\name{parse_pref}
\alias{parse_pref}
\title{parse xml node representing a prefecture}
\usage{
parse_pref(node)
}
\arguments{
\item{node}{xml node containing div.prefecture}
}
\description{
parse xml node representing a prefecture
}
