% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_list.R
\name{parse_station}
\alias{parse_station}
\title{parse xml node representing a station}
\usage{
parse_station(node)
}
\arguments{
\item{node}{xml node containg div.station without div.stmark}
}
\description{
parse xml node representing a station
}
