% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jma.R
\name{read_jma_monthly}
\alias{read_jma_monthly}
\title{read csv data from jma web site}
\usage{
read_jma_monthly(from, to, phpsessid)
}
\arguments{
\item{from}{Date}

\item{to}{Date}

\item{phpsessid}{php session id obtained by `get_phpsessionid()`}
}
\value{
csv data (character), can be processed with `readr::read_csv()`
}
\description{
read csv data from jma web site
}
