% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jma.R
\name{read_jma}
\alias{read_jma}
\title{read csv data from jma web site}
\usage{
read_jma(from, to, phpsessid)
}
\arguments{
\item{from}{Date}

\item{to}{Date}

\item{phpsessid}{php session id obtained by `get_phpsessionid()`}
}
\value{
csv data (character), can be processed with `readr::read_csv()`
}
\description{
read csv data from jma web site
}
\examples{
library("lubridate")
library("readr")

phpsessid <- get_phpsessionid()

# Site: Fukuchiyama, Kyoto
# Interval: from 2018-01-01 to 2018-02-01
# Data: Temperature and Precipitation

# lubridate::ymd()
f <- read_jma(ymd("2018-01-01"), ymd("2018-02-01"), phpsessid)

# parse csv using readr
col_names <- c("Datetime", "Temp", "Temp_exists",
"Temp_quality", "Rain", "Rain_exists", "Rain_qality")
col_types <- cols(col_datetime("\%Y/\%m/\%d \%H:\%M:\%S"),
col_double(), col_integer(), col_integer(),
col_double(), col_integer(), col_integer())
d <- read_csv(f, col_names=col_names, col_types=, skip=5)
head(d)

}
